Done my backend Work for Service `Support Ticket Management`
Done ✅:

- Attachments handling by building a middleware  (Thanks to `Multer`) 
- Added a way to only select entities which belong to the owner of the entity (in order to handle these broken authorization attacks that might happen/occur (Thanks to Mongoose's selection capabilities)  
- Comments CRUD operations , Now users can create/update/delete/read comments
- Ticket Create and read operations are Done (why not delete/update? because these operations will be handled in the `Customer service` that will be applicable only to `Support agents`
- for comments there will be two types of comments `Internal` and `External`, Internal will be only used by support agents to communicate each other, external is for both users support agents and customers, it makes sense that customers won't be able to view `internal` comments, so this case could be handled through the code 
- Added Security Headers by using the `Helmet` module 

TODO 📌: 
- refactor the code and implement that Authz vaildation logic on all entities (might create a middleware too)
- delete exif info from uploaded attachments (to make sure that privacy of our users aren't broken.

+++ services/ticket.js
@@ -2,31 +2,25 @@
 const TicketModel = require('../models/ticket');
 
 const AttachementModel = require('../models/attachment');
-const { model } = require('mongoose');
 
-module.exports.createTicket = async (ticketInfo) => {
-    try {
-        const { title, description, status, Owner } = ticketInfo;
-
+const jwt = require('jsonwebtoken');
+const CommentModel = require('../models/comment');
+module.exports.createTicket = async (title,description,userId) => {
         const ticket = new TicketModel({
             title: title,
-            description: description,
-            status: status,
-            Owner: Owner
+            body: description,
+            Owner: userId
         });
 
         await ticket.save();
-    }catch(err){
-        throw new Error("Error while creating ticket");
-    }
 }
 
 // TODO: Add validation to check if the user is the owner of the ticket
 module.exports.getTickets = async (userId) => {
     try {
         const tickets = await TicketModel.find({
-            Owner: userId
-        });
+            Owner: {$eq: userId}
+        }).select('-__v -Owner');
 
         return tickets;
     }catch(err){
@@ -49,51 +43,64 @@ module.exports.getTicket = async (ticketId) => {
 }
 
 // TODO: Add validation to check if the user is the owner of the ticket
-module.exports.AddComment = async (commentInfo) => {
-    try {
-        const { message, Owner, Ticket } = commentInfo;
-
-        const comment = new CommentModel({
-            comment: message,
-            Owner: Owner,
-            Ticket: Ticket
-        });
-
-        await comment.save();
-    }catch(err){
-        throw new Error("Error while adding comment");
-    }
+module.exports.AddComment = async (req,comment,userId,ticketId,type) => {
+        const userType = jwt.verify(req.headers.authorization.split(' ')[1], process.env.JWT_SECRET).userType;
+        let createdComment;
+        if(userType === 'customer'){
+            createdComment = new CommentModel({
+                comment: comment,
+                Owner: userId,
+                Ticket: ticketId
+            });
+        }else{
+            createdComment = new CommentModel({
+                comment: comment,
+                Owner: userId,
+                Ticket: ticketId,
+                Type: type
+            });
+        }
+        await createdComment.save();
 }
 
-// TOOD: Add validation to check if the user is the owner of the ticket
-module.exports.getComments = async (ticketId) => {
-    try {
-        const comments = await CommentModel.find({
-            Ticket: ticketId
-        });
+// DONE: instead of building a middleware to check if the user is the owner of the ticket, we can use the mongoose built in function ($eq) to check if the user is the owner of the ticket
+module.exports.getComments = async (req,ticketId) => {
+        const userRole = jwt.verify(req.headers.authorization.split(' ')[1], process.env.JWT_SECRET).userType;
+        const userId = jwt.verify(req.headers.authorization.split(' ')[1], process.env.JWT_SECRET).userId;
+        let comments;
+        if(userRole === 'supportagent'){
+            comments = await CommentModel.find({
+                Ticket: ticketId
+            }).select('-__v -Ticket');
+        }else{
+            comments = await CommentModel.find({
+                Ticket: ticketId,
+                Type: 'external',
+                Owner: {$eq: userId}
+            }).select('-__v -Ticket -Type');
+        }
         return comments;
-    }catch(err){
-        throw new Error("Error while getting comments");
-    }
 }
 
 // TODO: Add validation to check if the user is the owner of the ticket
-model.export.deleteComment = async (commentId) => {
-    try {
+module.exports.deleteComment = async (commentId) => {
         const comment = await CommentModel.findOne({
-            _id: commentId
+            _id: commentId,
+            userId
         });
-        await comment.delete();
-    }catch(err){
-        throw new Error("Error while deleting comment");
-    }
+        if(comment){
+            await comment.delete();
+        }else{
+            throw new Error("Comment not found");
+        }
 }
 
 // TODO: Add validation to check if the user is the owner of the ticket
-model.export.updateComment = async (commentId, message) => {
-    try {
+module.exports.updateComment = async (userId,commentId, message) => {
+    try{
         const comment = await CommentModel.findOne({
-            _id: commentId
+            _id: commentId,
+            Owner: {$eq: userId}
         });
         comment.comment = message;
         await comment.save();
@@ -102,3 +109,32 @@ model.export.updateComment = async (commentId, message) => {
     }
 }
 
+module.exports.addAttachment = async (ticketId, filename,filePath,userId) => {
+    const attachment = new AttachementModel({
+        name: filename,
+        url: filePath,
+        Owner: userId,
+        Ticket: ticketId
+    });
+    await attachment.save();
+}
+
+module.exports.findAttachments = async (userId,userType,ticketId) => {
+    try{
+    let attachments;
+    if(userType === 'customer'){
+        attachments = await AttachementModel.find({
+            Ticket: ticketId,
+            Owner: {$eq: userId}
+        }).select('-__v -Ticket -Owner -_id');
+    }else{
+        attachments = await AttachementModel.find({
+            Ticket: ticketId
+        }).select('-__v -Ticket');
+    }
+    return attachments;
+} catch(err){
+    throw new Error("Error while getting attachments");
+}
+
+}
